lab3/bot/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
greenlet==1.1.1
idna==3.3
install==1.3.4
numpy==1.21.4
opencv-python==4.5.3.56
pip==21.2.4
psycopg2-binary==2.9.2
pyTelegramBotAPI==4.2.0
requests==2.26.0
setuptools==58.0.4
SQLAlchemy==1.4.22
urllib3==1.26.7
wheel==0.37.0
